backend/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
python-docx
anthropic
openpyxl
python-multipart
pydantic
